Moved Files, updated paths, removed unecessary listener
